fuchsia: Use tiling field to check host access for dedicated memory.
A recent change (aosp/2164817) removed the virtualized "host-visible"
only heap and added host-visible bits back to device-local heaps.
Since Fuchsia relies on this feature to determine if a memory is used
for host access or device-only access, this makes it hard for goldfish
ICDs on Fuchsia to check if color buffer allocation is needed.

As a workaround, we now use the tiling field in VkImageCreateInfo
to check if host access is needed.

Test: scenic_pixeltest
Change-Id: If82beeb85c2f504a3cf976b64557a2e71509c6df