Support automatic NATT keepalive
This commit allows IKE Session to enable automatic NATT keepalive so
that IKE Session can choose the proper keepalive timer based on the
network environment. This feature allows IKE Session to cost
less power and at the same time remain the connection reliability.

To be able to support auto NATT keepalive, this commit also updates
the IKE Session to listen to the onCapabilitiesChanged callback
before migrating to a new default network.

Bug: 242231517
Test: atest FrameworksIkeTests(new tests), CtsIkeTestCases
Change-Id: I78d7852b4a0028611781d8c4e211578d8e6513ff