Autoupdate of 5 submodules
*bionic updated to c00d06867a7a27d8a610f81b0409a20fcc50db69
Merge "Use liblog_for_runtime_apex instead of liblog"
*system/apex updated to b98c0eee57d121214a771af1a664cf332631fa55
Use liblog_for_runtime_apex instead of liblog

liblog_for_runtime_apex is a static variant of liblog which is
explicitly marked as available to the runtime APEX. Any static
dependency to liblog from inside the runtime APEX is changed from liblog
to liblog_for_runtime_apex.

Previously, to support the need for using liblog inside the runtime
APEX, the entire (i.e. both static and shared variants) liblog module
was marked as available to the runtime APEX, although in reality only
the static variant of the library was needed there. This was not only
looking dirty, but also has caused a problem like b/241259844.

To fix this, liblog is separated into two parts. (1) liblog and (2)
liblog_for_runtime_apex. (1) no longer is available to the runtime APEX
and is intended to be depended on in most cases: either from the
non-updatable platform, or from other APEXes. (2) is a static library
which is explicitly marked as available to the runtime APEX and also
visible to certain modules that are included in the runtime APEX.

Bug: 241259844
Test: m and check that liblog depends on stub library of libc
Change-Id: I0347f180bca7bc5454e81ea3803979e11d922392

*system/core updated to f8053c661e5224a0723dc56624889e0cf0d55722
Merge "Use liblog_for_runtime_apex instead of liblog"
*system/linkerconfig updated to c0b06a3af679c6aff4932daa5757647c444da59a
Use liblog_for_runtime_apex instead of liblog

liblog_for_runtime_apex is a static variant of liblog which is
explicitly marked as available to the runtime APEX. Any static
dependency to liblog from inside the runtime APEX is changed from liblog
to liblog_for_runtime_apex.

Previously, to support the need for using liblog inside the runtime
APEX, the entire (i.e. both static and shared variants) liblog module
was marked as available to the runtime APEX, although in reality only
the static variant of the library was needed there. This was not only
looking dirty, but also has caused a problem like b/241259844.

To fix this, liblog is separated into two parts. (1) liblog and (2)
liblog_for_runtime_apex. (1) no longer is available to the runtime APEX
and is intended to be depended on in most cases: either from the
non-updatable platform, or from other APEXes. (2) is a static library
which is explicitly marked as available to the runtime APEX and also
visible to certain modules that are included in the runtime APEX.

Bug: 241259844
Test: m and check that liblog depends on stub library of libc
Change-Id: I7a95ce26147fca27cc7b6a4662d3617b3f0092ec

*system/logging updated to 00b275712694ef16192cbe7b99a58ef02b42527f
Merge "Use liblog_for_runtime_apex instead of liblog"