leaudio: Improve offloader behaviour
With this patch offloader is aware about number of CISes for the group
even whole group is not connected. Also it might be aware on which CISes are
connected based on the allocation value.

If flag
persist.bluetooth.leaudio.offloader.downmix_fallback is
set to false (default setup), the offloader will get complete list with allocation
already set to final allocation on stream start. There will be no updates of this
list during the streaming, but in this case offloader has no knowledge
about connection state  of CIS.

If persist.bluetooth.leaudio.offloader.downmix_fallback is
set to true, offloader will be notified about connection/disconnection
CIS using allocation value

Examples:
1) Group is banded headphones using one CIS to transmit two channels
   (left and right)

   In this case, offloader will get only one CIS with allocation set to
   stereo

2) Group is a earbuds which uses two CISes for single channel each, but
   only one earbud was connected.

   In this case offloader will get list of two CISes. The one which is
   connected will have allocation Stereo (left | right) and the one
   which is not connected will have allocation equal 0. Offloader is
   expected to mix riht and left

Bug: 231084798
Sponsor: @siyuanh
Test: atest BluetoothInstrumentationTests
Change-Id: Ic205348e45c4b8ecb3889f586d867675f0ce8931