Autoupdate of 2 submodules
*art updated to 754936451408e6dfe23a301f6b638d658cbc5d59
Report Odrefresh compilation times in milliseconds

Currently we are reporting the compilation times for the boot
classpath for primary/secondary architecture and system server
in seconds, but since the values we have observed so far are
4~8 seconds for the former and 8~16 for the latter, we might
benefit from having a finer granularity. Hence, this commit
is producing the compilation time in milliseconds (alongside
the previous values in seconds, for a transition period).

Bug: 243009337
Test: atest ArtGtestTargetChroot (especially art_odrefresh_tests)
Change-Id: Ib78a2e778d364653a5c29dd3430bf99274ca9a5a

*frameworks/proto_logging updated to 49f86d3b604f0c123a5460e49c1e8f0a0b6ec124
Merge "Add fields in OdrefreshReported for compilation times in milliseconds"