Autoupdate of 2 submodules
*frameworks/native updated to 549bf76e6084bfbb758f2463d114cf7b89f12413
Merge changes from topic "auto_delegator"

* changes:
  Add a delegate helper for AIDL Delegators
  Add lookupOrCreateWeak to ObjectManager

*system/tools/aidl updated to b87bca074a8003739e3120c6d3fd74e15c4bc17b
Make the cpp *Delegator smart!

If a binder object is an input argument to a call, the delegator now
wraps that binder object in its own delegator before sending it to the
delegate.
If a binder object is output argument / return value, the delegator now
wraps the binder object that was received from the delegate with its own
delegator before return the delegator to the caller.
This way, all calls to binder objects that are sent/received through
this delegator will still go back through this delegator.

This makes it possible to have a proxy process using this delegator
between a local binder interface and a remote binder interface with
binder RPC.

Bug: 220141324
Test: atest aidl_integration_test
Change-Id: Ic83d1d28647cfd00161b29e3da3ff21a7065c29e