Include art libraries in sdk snapshot for S+
Previously, the sdk snapshot would include all modules listed as
members for each build release even when the module specified that
it was not suitable for that release, by virtue of having a
min_sdk_version > target build release. A follow up change in
build/soong will exclude any module whose min_sdk_version > target
build release from the sdk snapshot for that build release.

Unfortunately, not every sdk member has a min_sdk_version specified. In
that case it defaults to use the sdk_version and if that is not
specified then it defaults to the "current" API level. Modules that are
also part of an updatable APEX are required to have a min_sdk_version
but other modules are not which means that those modules will not
appear in any snapshot other than the latest one.

This change adds a min_sdk_version: "S" property to the
art_cts_jvmti_test_library and expected_cts_outputs libraries to make
sure that they will continue to be included in the sdk snapshot for S+
once the follow up change is applied.

Bug: 240406019
Test: packages/modules/common/build/mainline_modules_sdks.sh
      # Ran the previous command with and without this change to make
      # sure that this change does not change the sdk snapshot
      # contents. Also ran the previous command with the follow up
      # change in build/soong to make sure that it does not change the
      # sdk snapshot either.
Change-Id: I32e61ec34cf83904c279a045e9754efd23890cd5