build_rootfs.sh: Fix serial gettys on rockpi
The copy/pasted code from cuttlefish referred to the wrong serial port
(ttyS0 instead of ttyS2) and to a virtual platform only device, hvc1.
Remove hvc1 and update to ttyS2. This has no real impact as getty
already spawned on tty1, and because we don't configure any graphical
output this is already the serial console.

Change-Id: Idb01f1a0b5ff7b38887957d7332685c3f42c20da