Autoupdate of 2 submodules
*frameworks/proto_logging updated to ce8c0213397139eef2898e14e799e587ac5e4a65
Replace established atom required fields with optional

When move the new atom(aosp/2182097) to google3, It got the
error as "Server.validated is not an allowed required field".
According to go/required-fields-harmful, we should use optional
instead of required.

Test: Run statsd_testdrive
Bug: 241788170
Change-Id: I45c7107d0059e3de744e1b359e7f23a6bdc7dbd6

*packages/modules/DnsResolver updated to 05668a1347980c56e7c354cd4b9edbd52954af43
Replace established atom required fields with optional

When move the new atom(aosp/2182097) to google3, It got the
error as "Server.validated is not an allowed required field".
According to go/required-fields-harmful, we should use optional
instead of required.

Test: Run statsd_testdrive
Bug: 241788170
Change-Id: I1de03a7d0b0f36f446aa4ffb3e1b4d6e1d659eed