goldfish-opengl: add stub VirtGpu layer for Fuchsia/host builds
This adds an implementation of the platform layer to Fuchsia and
the host build.  The goal is to prevent further #ifdefs in code
and try to localize platform-specific code with build systems,
as is standard practice.

Long-term, one may split out a platform/fuchsia directory with
actual implementations.  For example, if the goldfish address space
and sync drivers are eventually going to be replaced by virtgpu
in AEMU, then it makes sense for Fuchsia to adopt a
fuchsia.hardware.virtgpu rather than fuchsia.hardware.goldfish.
That would make running Fuchsia emulator graphics possible on other
VMs that implement virtio, such as crosvm.

In addition, it should also be possible to split out a
platform/host directory too if needed.  Currently, the host build
unit tests have a heavy dependence on Goldfish and related HALs
(for example, the vulkanhal_unittests use Goldfish gralloc, which
is being deprecated).  As AEMU goes to virtgpu, it's worth thinking
about how to get the same coverage going forward.

BUG=233803018
TEST=fx build, ./rebuild.sh

Change-Id: Iec7ed92f9bf7d9c895c68652373a580e97014798