Leaudio: Fix one earbud no playing when both connected at once
Issue could happen when two devices where connected and attach to the
stream just after enabling Bluetooth. What is happening is that first
connected device start its way to Streaming state but it is not there.
In this time if second device connects, it tries to attach to the stream
but it is not allowed because group is not yet in streaming state.
In such a case we should wait a bit.

Bug: 240892829
Bug: 231084798
Test: atest BluetoothInstrumentationTests
Tag: #feature

Change-Id: I93f5ea912450cf18ad3d419e48c29a7f1acbea16