Set CaptivePortalLoginTests timeout to 10min
The test-timeout needs to be set via annotation on each test, as the
test infrastructure is already overriding the run-wide
AndroidJUnitTest:test-timeout setting.

The test is currently failing in presubmit runs (and not in postsubmit
runs of the same test) as the first method run of the class takes >1min,
which is the default timeout.

Bug: 244275469
Test: atest CaptivePortalTests
Change-Id: I44c69739357f17d36a4d246fd531641134a8fa1b