Create local pip repository
This updates the embedded test to install all the package from a local
repository, which is included here.

This will make sure that all required packages are available, and that
no outgoing network connections will be made.

It updates the documentation to describe the steps to take when wheels
needed for your platform are missing.

Bug: 244438112
Change-Id: Iae67a61d51eea63be5f3e04b612641216432a656