Autoupdate of 2 submodules
*art updated to 57e81e257bab58e8d3e2982a68d2b31ff306ecd7
Don't copy prebuilt implementation libs that should be built from
source.

See https://r.android.com/2198690 for details.

Also added a note on how standalone-apex-files is broken. It'd be a lot
of work to make that approach safe. Instead refer to b/129332183, which
is about moving Golem to chroot, which would make this whole thing
obsolete.

Test: art/tools/run-gtests.sh
      art/test/testrunner/testrunner.py --target --64
  in the ART chroot setup.
Bug: 129332183
Bug: 243935575
Change-Id: I43c6f8b6315bc9cdd0d0e7a6bf57ce4ca3f5f60e

*prebuilts/runtime updated to a040c5e282b0d5f6ba73a986625eff1a4bea792d
Merge "Remove some platform prebuilts that are built from source in master-art."