Skip pKVM boot time benchmark on _pkvm builds
The recently introduced testBootEnableAndDisablePKVM fails on _pkvm
builds because the pkvm_enabler service checks if pKVM is enabled and if
not, sets the corresponding flag in the misc partition and reboots the
device. The test overrides this config in fastboot, resulting in a
bootloop and making lab devices unavailable. Skip the test if it detects
that the flag in misc has been set.

Bug: 236799228
Bug: 244769139
Test: atest AVFHostTestCases on oriole_pkvm-userdebug
Change-Id: I20435fa3619ae428c167de4195ef1acf40261057