Use cached static available modes in getAllChipInfo()
Currently we allow creation of driver-specified ifaces after the driver
is ready, and driver availableModes are returned by the current chip
info.

However, driver-specified ifaces may be requested before the driver is
ready, but the availableModes returned by the chip will still be the
default modes. Thus we should populate the current chip info with the
cached static modes to allow creating driver-specified ifaces before
the driver is ready.

Bug: 239532073
Test: atest HalDeviceManagerTest

Change-Id: Ied51391053bad893623609e4871f18bdaf57e89f