Don't include hidden classes in the class graph
When a hidden/private class is found while constructing the graph, this keeps traversing the inheritance chain but leaves the class out of the graph. So if A extends B and B extends C but B is hidden, C is listed as the direct parent of A and A as the direct child of C.

This doesn't deal with the inherited methods section yet.

Test: Added integration tests.
Bug: 243324828
Change-Id: I523f5ea850e2e78b9bef8af63a3859df3621e1b6