Use heck-0.3.3
Structopt is EOL and will not update to 0.4
The plan for migrating off is to delete structopt after migrating
clients to clap + clap-derive

"tests" is removed from cargo2android.json to avoid a bug in
cargo2android.py which will currently generate a test target when there
are 0 tests.

As this project is EOL and has no tests, this should not cause trouble.

Test: mm
Change-Id: I31824494e12bd6fd66a3effe4bf756a896815045