Modify aidl tracing code gen for Java
Enabling aidl#generate_traces in aosp/2190022 increases framework.jar
size by ~3%. This is primarily due to the increase in transaction name
strings in the dex. The real issue here is that the dex size increases
proportional to the number of methods vs the number of interfaces

To fix this, we take advantage of the already generated txn name strings
in the framework.jar aidl as part of the enabled
aidl#'generate_get_transaction_name'. This tag creates a method
getDefaultTransactionName(txnCode) with a switch statement to convert
txnCode -> txnName. We can then emit traceBegin code once per Stub
interface that looks like:

boolean onTransact(int code...) {
  try {
    ...
    traceBegin(getDefaultTransactionName(code))
    ...
  } finally {
    traceEnd(...)
  }
}

instead of emitting a traceBegin per txn case statement.

Unfortunately, this approach does not extend to the Proxy interface
because each txn exposes an individual method and there's no single
entry point to emit a traceBegin like the Stub#onTransact. Hence we
skip tracing the Proxy side. This means we don't get any 'client'
aidl tracing for Java aidl interfaces but this was arguably unnecessary
anyways and might be worth considering dropping that for the Cpp backend

Test: Manual
Bug: 161393989

Change-Id: Ib78d82ceea41d13e047337907a1f49f9f1615a86