Autoupdate of 5 submodules
*external/rust/crates/argh_derive updated to 373aa013134e5f330ea9e744d479f591270e71f0
Use libheck-0.3.3

Upstream has not updated to 0.4 yet.
As it is maintained by Fuchsia, we expect they will.

Test: mm
Change-Id: I77be86f783d4c34cafb52742b2a40a196d5cba6f

*external/rust/crates/clap updated to 378471649a2ed11aecc139e3cca33df47693c544
Enable derive feature to use clap-derive

Test: mm
Change-Id: Ic95fb7a0b72d10dfdfc0596898ef4da156efe9ca

*external/rust/crates/clap_derive updated to 45465bece3bec57b602d16a0ce09115e6aafb6e6
Initial Android.bp generation

Test: mm
Change-Id: I6b31b8fbc0d4362864d732aa093f3cf584bd11a8

*external/rust/crates/heck updated to 8e19def4203f1a4ca7cf553c8b74fa7608743c22
Version split - 0.3.3 -> legacy, 0.4 -> current

Test: mm
Change-Id: Iaa94208ce0b7060f5280e28183e4158e34c403dd

*external/rust/crates/structopt-derive updated to 960abda7c6acb3da377f31401cd7e70991bf1a74
Use heck-0.3.3

Structopt is EOL and will not update to 0.4
The plan for migrating off is to delete structopt after migrating
clients to clap + clap-derive

"tests" is removed from cargo2android.json to avoid a bug in
cargo2android.py which will currently generate a test target when there
are 0 tests.

As this project is EOL and has no tests, this should not cause trouble.

Test: mm
Change-Id: I31824494e12bd6fd66a3effe4bf756a896815045