task_profiles.json: Switch to the libprocessgroup version
Emulations should be as close as possible to Android running on a phone.
Hence switch from the Goldfish version of task_profiles.json to the
libprocessgroup version of task_profiles.json. This change activates the
I/O scheduler settings that are present in the libprocessgroup version
of task_profiles.json but not in the Goldfish version.

Bug: 227389363
Test: Verified that logging in with adb to the Trusty emulator still works.
Change-Id: I710f1831dbdf1f7116f58b37955e65f52af5bf15
Signed-off-by: Bart Van Assche <bvanassche@google.com>