Override CF's wifi interface name.
Cuttlefish has its wifi interface name set to "eth2" because it makes
also use of other interfaces (i.e. eth0, eth1). In Trout there is only
one interface in use so there is no need to handle the other ones.
Because of that fact, Wifi has to be set up on "eth0".

Bug: 246273743
Test: Build and connect to WiFi
Change-Id: Ic12fa419261b88ecf4b2783db16eaa251ba11b6f