Add Listening Funcionality to AIDL Analyzers
Added the ability to listen to a service in real time,
rather than needing to first record and then inspect after.
Upon starting the listening process with an instance and
interface, transactions will be printed until the user
hits CTRL-C for a ctrl_c_limit times. A file is also created
with all of the transactions that were listened to.

Bug: 242226631
Test: atest aidl_integration_test
Change-Id: I433b8a0911bb1871cf00aaf7a7e14a7da23671ca