Autoupdate of 2 submodules
*frameworks/libs/net updated to 32007723a052674da2744adfe42df3e90433afac
RaResponder: add support for NS/NA for tests

Add support to respond to neighbor solicitations in addition to router
solicitations. This is useful for tests that need to send IPv6 traffic
through a tap interface. I have previously experimented with the target
link-layer address option in the RA, but it did not seem to prevent a
subsequent NS.

As a TODO, this class should be renamed to NdResponder (and possibly
rewritten in Kotlin).

Test: test only change
Bug: 235559605
Change-Id: I33bac507082fa89d1a5ae02cbe671ce7f5f09b07

*packages/modules/Connectivity updated to 4fdc5fda4aa9c6253148d408d78c478d31783347
DO NOT MERGE: DscpPolicy: remove post-T dependencies from test

The DscpPolicyTest needs to be merged into CTS13 R2, so it cannot
always use APIs that were introduced after the T cutoff.

Run the test when the new APIs are available, but skip it otherwise as
the feature is known to be broken unless the device is using a newer
connectivity module.

DO NOT MERGE: should not be used for U+ tests.

Test: atest DscpPolicyTest
Bug: 243891394
Bug: 243120975
Change-Id: I86eeb7a008c6e299fdf7cb709844bc336f09bdfa