Add an option to save apk files to output artifacts based on test results
Previously when we verify an app test test result, we need to download the apk manually from a source. This is not convenient, and we may download an with different version which make us unable to reproduce the test results. Saving the apks allows to use the exact test input file in order to easily reproduce the test results.

Test: unit test
Change-Id: I750a362b5e74278afdbb3ed8e3c104fa991302e8