Rebuild crosvm with fixed minijail
Previous builds were not using the libminijail.so prebuilt, and minijail
was being built incorrectly using host glibc features in a non-hermetic
way. Fix the Cargo build of minijail-sys to use a shared library and fix
the shared library to use the same prebuilt intermediates that
Android.bp uses.

This restores proper functionality of the crosvm sandbox while still
only requiring glibc 2.8.

Bug: 246436720
Change-Id: Idb5d07e50b364529f84b6cb0988918af6153948e