Autoupdate of 3 submodules
*external/adt-infra updated to f988128f9443ee6faf56f43f02fccc44040bfe06
Make the build hermetic

This makes sure we use the shipped python interpreter, vs. the one that
is on the build bots.

Includes updated documentation outlining what the script is doing.

Bug: 245390993
Change-Id: Ia1dbc65de027d3dfb1c9c9bf3028395535e4385c

*external/qemu updated to 06715e62de7d4004817d66314f7067fb50b93334
Merge "Use the internal python toolchain" into emu-master-dev
*tools/buildSrc updated to 5da63ce17ae211f37d15fdce4dcb409929efe9d6
Enable the build scripts to use AOSP Python3

This makes sure that the build scripts can use the internal Python3
interpreter.

Bug: 245390993
Change-Id: I34e4f969d59509448402cd97c948ddc82e4340f8