Autoupdate of 2 submodules
*build/soong updated to 7113b19be803d54abca058b9d9384bee0625b206
Add a host tool to create an APEX bundle

Even though the soong build system generates a bundle module for an
apex, the bundle module itself should be further processed to be an
AppBundle (.aab) file which can be uploaded to Play.

This script fills the gap by invoking bundletool to create an
AppBundle (.aab) file out of soong-built bundle module for an apex.
(Note: uploading APEX bundle (.aab) to Play is not supported yet.)

You can create an .aab file by:
- TARGET_BUILD_APPS={apex name} m dist
- m build-apex-bundle
- build-apex-bundle --output out.aab out/dist/{apex name}-base.zip

For now it creates a single-ABI APEX bundle. In the future it can be
extended to support multiple-ABI APEX bundles.

Bug: 236673372
Test: m build-apex-bundle
Test: TARGET_BUILD_APPS=com.google.cf.bt m dist
Test: build-apex-bundle --output bt.aab out/dist/com.google.cf.bt-base.zip
Change-Id: Id321efcd42c0fe60294a8348047c9ebbf7acf391

*prebuilts/bundletool updated to 399c1462a9c0b3ccb0fafe7d791072b062f219b3
Add bundletool java_binary_host module

This is for convenience.

Bug: 236673372
Test: m bundletool
Change-Id: I8f6f7eef7004ba126b520fa2ed0870eca76de97c