Update TEST_MAPPING with new key names
Because aosp-master-plus-rust is downstream of aosp-master, if we use
the "presubmit" or "postsubmit" `test-group`s in our ATPv2 test mapping
definitions it will trigger tests for any files changed in AOSP, not
just those relevant to the Rust build.  To address this issue we will
use new keys (rust-toolchain-presubmit and rust-toolchain-postsubmit)
that will only be used in toolchain/android_rust and toolchain/rustc.
As a result, only the tests defined in those directories will be run by
aosp-master-plus-rust's test mapping target.

Test: Can't be tested until it's submitted
Bug: 166837771
Change-Id: I867874585c98d4f3f2cd9f8e122ad679c73b034b