Fix builds with an absolute OUT_DIR
protoc assumes an include path of "." if none is specified, which is
the top of the source tree.  When OUT_DIR is an absolute path,
the path to the genrule sandbox is not inside the top of the source
tree, and protoc can't map the sepolicy.proto file to an include
directory.  Add the directory that contains sepolicy.proto to the
include directories.

Bug: 243196682
Test: OUT_DIR=/tmp/out m /tmp/out/soong/.intermediates/packages/modules/SEPolicy/API-level-policy/33/SEPolicy-33.pb.gen/gen/SEPolicy-33.pb.gen
Change-Id: I203e4424bdaeae57c03d172c97f9bfd91426106c