Update Test924 to filter for messages related to test thread
Test924 tests that the thread messages are working as expected.
To test this, it constructs a thread, yields and sleeps and then
starts the thread. It expects that there are no events between the
construction of the thread and starting the thread. Though it is not
always guaranteed that there won't be activity corresponding to some
other thread. To prevent corner cases, filter out messages corresponding
to other threads.

Bug: 246688031
Change-Id: I36445d8df87c748b472b21b1a49e11b92907fea1