ANDROID: run_c2a.sh: explicitly pin rust version
We should be getting a consistent version because of the rust-toolchain
file, but anecdotally we are seeing mysterious differences in behavior
when running C2A. Maybe this will help.

An extra benefit is that we don't need to explicitly specify the
cargo-bin param. We can't accidentally forget it, which then uses the
prebuilt binaries in AOSP that don't completely work for us and lead to
other differences.

Bug: 246990922
Test: all2android.sh
Change-Id: Ib8ef125ae08c20706e95e6ca94545f6fb7a02b92