Reland: Support disallowing concurrent sessions
Previous: aosp/2203568

A custom data source class can have
`static constexpr bool kSupportsMultipleInstances = false;`
member to indicate its support for multiple instances. It is true by
default.

When this flag is false, we cannot have multiple instances of that data
sources. When a data source is already active and if we attempt
to start another instance of that data source (via another tracing
session), it will fail.

We have also added a flag named `support_concurrent_sessions` in
TracingInitArgs. It is true by default. When its is set to false,
it overrides `DataSource::kSupportsMultipleInstances` for all
the data sources.

Bug: b/245118150
Change-Id: I4f62c9aaf4ca2d4c9dad0a6302a7357794619988