Remove support of @EnforcePermission on classes
The intent of the @EnforcePermission annotation is to visually indicate
that permissions have already been verified. When applied to a class,
the visual indication is far from the actual implementation (method).

Only accept the annotation on methods instead of any type. This
annotation has recently been introduced for platform services only.
There are no instances of such class annotation in AOSP.

Test: build
Change-Id: I2d6a747cc41badf5ce8788986a4434ed7add0790