Temporary workaround to set platform flag.
This CL interprets a min-api of 10000 to imply that the build is
a platfrom build. This is needed to enable optimizations that
assume the runtime bootclass path is that of the current platform.
This workaround must be removed prior to branching platform and
changing the default min-api a different value than 10000.

Bug: 249230932
Bug: 246905006
Test: n/a
Change-Id: If837967c83268a27f9ede42c4216b72736b93139