Autoupdate of 5 submodules
*art updated to e6e6bee183bf69cf313a3b942a84cd18cbaae31c
Annotate symbols correctly

Currently, tagging a symbol with #apex (# systemapi or # llndk) is not
required when the symbol is in a non-NDK library. However, this is
considered dangerous because such a symbol will automatically be
promoted to NDK APIs when the library is promoted to an NDK library.
When that happens, the native API council won't be able to notice the
promotion because promoting a non-NDK library into an NDK library
doesn't require an update of the map.txt file, but Android.bp only.

To prevent that, we should mandate those tags for Mainline APIs
regardless of whether the library the API belongs to is an NDK library
or not.

Upcoming changes in build/soong will enforce this. This change is to
prepare for the enforcement.

Note that this is a build-time only change. There's no behavior change
at runtime.

Bug: 184712170
Test: m
Change-Id: I18d9722177cd440b939d090b9b653405409a8b57

*external/icu updated to 6db38a9d4ece8697550f70d4997a2eb7f8a361b6
Annotate symbols correctly

Currently, tagging a symbol with #apex (# systemapi or # llndk) is not
required when the symbol is in a non-NDK library. However, this is
considered dangerous because such a symbol will automatically be
promoted to NDK APIs when the library is promoted to an NDK library.
When that happens, the native API council won't be able to notice the
promotion because promoting a non-NDK library into an NDK library
doesn't require an update of the map.txt file, but Android.bp only.

To prevent that, we should mandate those tags for Mainline APIs
regardless of whether the library the API belongs to is an NDK library
or not.

Upcoming changes in build/soong will enforce this. This change is to
prepare for the enforcement.

Note that this is a build-time only change. There's no behavior change
at runtime.

Bug: 184712170
Test: m
Change-Id: I62b47e59688ecb7d2eb6da45910863e8b8caff29

*external/selinux updated to 5f3411ec65b3733a405011027f20e1bad8ca185f
Merge "Annotate symbols correctly"
*libcore updated to a6de061775725208672be13ced2cd6a46b9ceb44
Merge "Annotate symbols correctly"
*system/core updated to b2d8315f10b242daf604c24a5f1e8007b13f86fe
Merge "Annotate symbols correctly"