Autoupdate of 4 submodules
*external/selinux updated to 5f3411ec65b3733a405011027f20e1bad8ca185f
Merge "Annotate symbols correctly"
*packages/modules/DnsResolver updated to 7eeed85c764860d6aedab0562814a6422d6ff0d1
Merge "Annotate symbols correctly"
*packages/modules/adb updated to 08b703261f5d9a044bb9cb9234a332260736af39
Annotate symbols correctly

Currently, tagging a symbol with #apex (# systemapi or # llndk) is not
required when the symbol is in a non-NDK library. However, this is
considered dangerous because such a symbol will automatically be
promoted to NDK APIs when the library is promoted to an NDK library.
When that happens, the native API council won't be able to notice the
promotion because promoting a non-NDK library into an NDK library
doesn't require an update of the map.txt file, but Android.bp only.

To prevent that, we should mandate those tags for Mainline APIs
regardless of whether the library the API belongs to is an NDK library
or not.

Upcoming changes in build/soong will enforce this. This change is to
prepare for the enforcement.

Note that this is a build-time only change. There's no behavior change
at runtime.

Bug: 184712170
Test: m
Change-Id: Iae8516ba278f0a17bd50eeba94dd7596c8d9aa32

*system/core updated to b2d8315f10b242daf604c24a5f1e8007b13f86fe
Merge "Annotate symbols correctly"