Correctly handle non-invokable methods in InitializeMethodsCode
Non-invokable methods were handled correctly when there is no
instrumentation but it was not obvious or explicitly handled. For ex: we
could still install AOT code but we never pass AOT code for
non-invokable methods. When instrumentation was enabled, non-invokable
methods weren't handled correctly. This CL makes the handling explicit
by initializing the entrypoint to QuickToInterpreterEntryBridge.

Bug: 206029744
Test: art/testrunner.py -t 2240-tracing-non-invokable-method
Change-Id: Idf2e95e29791d0e4244afa5dfe36e5b78649853c