Rounds up image size for add_hashtree_footer
`avbtool add_hashtree_footer` is supposed to round image size
to be a multiple of block size when --partition_size is provided.

However, when invoking append_raw() to round up image size, an
assertion error will occur because the length of the padding data
is not a multiple of block size.

Changing append_raw() to allow skip checking the input size,
so we can round up the image size.

Bug: 249402799
Test: run libavb_host_unittest
Test: run libavb_host_unittest_sha
Change-Id: Icd08a2b0f808f6a9bb5e4eab18f8ac6180e7d667