Add a Bazel test suite target
Create a target to package artifacts and configuration for a TF suite
that can run the Bazel tests contained in the Atest-generated workspace
archive.

The suite does not use standard Make-based suite packaging rules since
there are no test modules to tag (e.g. "general-tests") as being part
of the suite. Instead the suite contains a single test module which
delegates test execution to Bazel.

Test: m bazel-test-suite
Bug: 202790559
Change-Id: I767faa258d13cdbdbcde741f48fad2a05659d78e