onAvailable should check if network has changed
This commit updates IkeDefaultNetworkCallback#onAvailable to
only trigger onUnderlyingNetworkUpdated when the input network is not
the same with the mCurrrNetwork.

Without this fix, IKE Session setup with the default network will always
fail. The reason is the onAvailable will always be called immediately
after the callback is registered. Although this onAvailable call does
not mean the underlying network has changed, the IKE Session will
treat it as network change and will terminate the session because
the mobility is not enabled.

Bug: 250689403
Test: atest FrameworksIkeTests(new tests), CtsIkeTestCases
Test: manually tested
Change-Id: I5bc5d8549e3e72c4e8b5ed71de25c27f9bb38402