Add libgolemtiagent.so to set breakpoints on golem benchmarks
We are updating debuggable-cc configuration in golem benchmarks to
measure the performance of debugger. As a first step, we are adding
a breakpoint to measure the overhead of instrumentation related code.
To achieve this we need a tiagent to be able to set a breakpoint. This
CL adds a minimal ti agent that could be used in golem benchmarks.

The changes in this CL:
1. Adds Agent_OnLoad callback which installs a listener for VM_INIT
   event.
2. On the VM_INIT callback we set a breakpoint on Thread.stop which is a
   deprecated method and not expected to be used much.
3. Adds rules to build libgolemtiagent.so for build-art-host-golem
   target.

Bug: 232212577
Test: m build-art-host-golem
Change-Id: I2da3770ff85edb46da64d09ec0047ec6e7564fd0