Autoupdate of 2 submodules
*external/libgav1 updated to 8c0ea6aa4c3efad668e53ebcf3fd569f3f049851
Move libgav1/* to *

This is needed to get the folder structure to be in sync with main
branch of upstream project.

Except the changes in Android.bp (as per the new folder structure),
all other changes in this CL are due to files being moved without
any changes in the code.

Bug: 211157723
Test: Builds
Change-Id: I96af139af664c32da49f15e19ccc592750b46e3e

*frameworks/av updated to 4aec31d0be9d61f29933fbc88044e6f678f376b0
Merge "C2SoftGav1Dec: Update include paths"