Disable CF's rename_eth0 one-shot service.
rename_eth0 service is used in CF [1] and, as its name says, it renames
eth0 interface. In trout this operation is done in setup_wifi [2].
Disable it in CF to avoid errors.

Cuttlefish has its wifi interface name set to "eth2" because it makes
also use of other interfaces (i.e. eth0, eth1). In Trout there is only
one interface in use so there is no need to handle the other ones.
Because of that fact, Wifi has to be set up on "eth0".

[1]: https://android-review.googlesource.com/c/device/google/cuttlefish/+/2214685/3/shared/config/init.vendor.rc#58
[2]: https://android.googlesource.com/device/google/cuttlefish/+/a376692d779395f63ba84d3733ed12f959669e6d/guest/commands/setup_wifi/main.cpp#142

Bug: 246273743
Test: Build and connect to WiFi
Change-Id: I9f865273fbd77fec48472e2a2f4fdd82b4976cb2