Adding More Wait Time in ShellSubscriber CTS Test
Adding wait time between when subscriptions are initiated and when a
pushed atom event occurs to make sure the subscriptions are active when
the event occurs. Adding wait time between when pushed atom event occurs
and subscription processes are killed to ensure that subscriptions
process the pushed event before being killed. Adding wait time between
when processes are killed and new processes are initiated to make sure
that max subscription capacity isn't unintentionally hit. Adding prints
to help debug in the future.

Bug: 251470843
Test: ShellSubscriberTest

Change-Id: I218dcd59b5e51273ddecbd9e49c3b01ee9f5381f