Replace "rkp only" semantic with explicit blocking calls
It turns out to be simpler to replace the relatively complex "rkp only"
registration with an explicitly blocking call to fetch keys.

This puts more control of whether to block in the hands of the client.
For example, even rkp-only clients may want non-blocking calls so they
have control over retry loops.

Test: It Builds
Change-Id: I60ccead036af2e5b4be51ff86bb7fc329cf60c24