Do not shut down IwlanDataServiceHandler in deInitNetworkCallback
When phone app crashes, IwlanDataService.onUnbind() will be triggered and Iwlan will force close all tunnels. Also, as part of DataService.onUnbind(), all IwlanDataServiceProviders will be closed and IwlanDataServiceHandler will be shut down in deInitNetworkCallBack().

When EpdgTunnelManager closed the tunnel and report to IwlanDataService via IwlanTunnelCallback, as the handler is already shut down, Iwlan was not able to process the event and raise NullPointerException.

Fixing this issue by not shutting down IwlanDataServiceHandler.

Bug: 245701042
Test: atest IwlanDataServiceTest, verified in live test by crashing phone app with adb command
Change-Id: Idfed9a4e67a1b8d0371a3afd7a4a70df8af13f40