ui: Remove commented-out screenshot test
Subjective opinion: it's not useful to keep the commented-out test in
the code. The TODO has been added a year ago and so far has not been
addressed. Keeping the dead code just adds extra stuff to scroll through
when looking at screenshot tests code.

In order to preserve the code, I've copied it to the comment in the bug;
if someone would address the bug, they would have it as a starting
point.

Bug: b/198431341
Change-Id: I7ee4d3f9a4c6490b997ae8cba0cea736a2ead858