AIDL: gen_rpc by default
Always use 'Parcel.obtain(IBinder)' when available. This is required in order
to use RPC, and we want code to always work in both situations.

One downside of this is that RPC is no longer explicit (someone may be confused
if they see a new SDK version target has RPC magically working, but they fail
with a runtime error). However, since we want to move away from having a
difference at all, we say:
- the runtime error is good enough
- we still have 'gen_rpc' as-in 'require_rpc' (won't rename due to source cost)
  which can be enabled and will give the clear min SDK version error

Bug: 248249899
Test: goldens
Change-Id: I29f6a5f2d30f9a27b0eba9709868a4c1ed4e84d6