attempt to wipe digital car keys when clearing userdata from fastbootd
A WipeKeys function is optionally compiled into the recovery UI on
Pixels where digital car keys are supported. This change looks for that
function and, if found, uses it to clear the keys during a userdata
wipe.

Includes are reordered based on repohook requirements.

Bug: 203234558
Test: adb reboot fastboot && fastboot -w; then check for key after reboot
Signed-off-by: Jeffrey Carlyle <jcarlyle@google.com>
Change-Id: I46a58f7d521fdefd0675be0b4180bbefdee2e660
Merged-In: I46a58f7d521fdefd0675be0b4180bbefdee2e660