MediaPreparer: Use media-folder-name in path for bbb_* files
Added a parameter "use-legacy-folder-structure" (default: true),
which is used to control folder structure for bbb_* files used in
mediastress tests.

When "use-legacy-folder-structure" is set to false, instead of
/sdcard/test/bbb_*, /sdcard/test/<media-folder-name>/bbb_*  is used.

This approach allows for any potential future updates to the contents
of the zip file and the extra parameter "use-legacy-folder-structure"
maintains backward compatibility when tradefed prebuilt binaries with
this change are ported but the mediastress changes in cts aren't.

CtsMediaStressTestcases which uses these bbb_* files is being updated
to use its own zip file which would be version controlled and can
potentially change frequently.

Bug: 230128098
Test: atest CtsMediaStressTestCases
Test: atest --no-bazel-mode \
 com.android.compatibility.common.tradefed.targetprep.MediaPreparerTest

Change-Id: I5c3c07db656ca2de6e567f66ca4b885382184a03