Remove mocking of MimetypeMap and mockito-extended
In production code, the file open intent always used the mimetype
detected when starting downloads, since:
 - For directly-open files, the "done" notification is not shown, and
   CaptivePortalLoginActivity already used Intent#setDataAndType
 - For other downloads, the download was done to a FileProvider-managed
   file created through ACTION_CREATE_DOCUMENT, with the mimetype set to
   the download task mimetype. ACTION_CREATE_DOCUMENT is documented to
   provide a content:// URI with the specified mimetype, so
   Intent#resolveTypeIfNeeded would just return what was set previously.

So MimetypeMap#getMimeTypeFromExtension was only used to resolve a
mimetype in tests, where download is done to a test-created destination
file that has no pre-set mimetype.

Set the mimetype explicitly in makeDoneNotification instead of relying
on resolveTypeIfNeeded, which would return the same value in production
code anyway. This avoids relying on the mimetype of the file URI, so
tests can expect that the mimetype of the download task and the intent
will be the same, regardless of the destination file used in tests.

With this, CaptivePortalLoginTests do not depend on mockito-extended
anymore, so it can be removed.

Bug: 243315116
Test: atest
Change-Id: Iabb2f2599a682baabb530d1dab831713c8fefd11