Remove session_manager_lib.
This change removes the session_manager library. It also removes
re-starting a session during clean_up of the wlan_policy_controller,
which is not needed by our tests with the eng_arrested builds. We keep
the session stop logic in wlan_policy_controller to support other
developers potentially using different builds that need to stop the
session.

Fix: 249840399
Bug: 249840399
Test: act.py [...] -tc FlashTest HiddenNetworksTest
Change-Id: Idf8942c603bf51a12258d746f0ea8b6ae770dfdd