goldfish-opengl: create multiple timelines even when no capset is specified
Multiple timlines are only requested by the VK host connection, since
it's the only part of gfxstream that uses the feature.  However, it
is possible an app creates a GLES connection before a VK connection,
and since there's a single virtgpu device, that causes command
submission failures.

It wouldn't happen when guest ANGLE is used.

For now, just request multiple timelines always.  Command submissions
will go on the global timeline still for GLES, since the the GLES
encoder doesn't specify the kRingIdx flag.

BUG=254065847
TEST=run 3dmark slingshot extreme, observe no cmd submission failures

Change-Id: I4395217032c5e9918018cecd0a0e0c0a3922464d