Add source dimension to TlsHandshakeReported.
Reason for cherry-pick to AOSP: Metrics are logged from
an AOSP-first Mainline module.

There are potentially multiple versions of Conscrypt on
a device.  Mainline, GMS, or included with an app by its
developer.  A single app could conceivably contain and use
all three, and all three may be different versions.

This change adds a source dimension to allow us to differentiate
which version is used and monitor release quality of each
separately.  Until the logging code in GMS and Mainline is
updated, everything will appear to be UNKNOWN.

Obviously this will require us to change our metric definitions
to pick out the different sources but as these metrics are
still in modulefood this shouldn't be problematic.

No privacy impact as there is no way to identify the user,
device or app from the flavour of Conscrypt in use. So assuming
this is covered by our existing PDD?

For storage, this Atoms is still in modulefood so assuming a
very small increase in storage required.

Test: m
Change-Id: I7b995531ee73a71925077d1498f63559e70fc5af
Merged-In: I7b995531ee73a71925077d1498f63559e70fc5af
(cherry picked from commit f930fead4b7bc3db30f8aba3fa9e596aa0ed6588)