bpf_base_test - unconditionally run TestSocketFilterSetUp
We require 4.14 kernel so the condition is always true, and
thus this test is never actually skipped.

Additionally this test is outright incorrect, because BpfHandler::initMaps() requires the uid permission map open
to succeed, and furthermore BpfHandler::hasUpdateDeviceStatsPermission
requires it for things to even work, as this is used as part
of tagSocket() permission checks.

(side note: it's ugly that this netd test is reaching into
mainline stuff... but fixing that is a different TODO)

Test: TreeHugger
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I4ea401e8bd73f9c3fce4498582dbcdf247a9453b