Remove timestamps from api diff files by using labels.
The api diff files contain the timestamp of files which are being
diffed. This results in different diffs getting generated everytime
even if all the APIs are the same.
In order to generate same diffs, timestamps have been removed by
using the label flag while performing the diff operation.

Bug: 238757360
Test: The builds passed and the api diff files does not contain
timestamps.

Change-Id: If894f5a6632542b566eb0657d22d4c4b2f6afa83